refactor(broker): rename internal/oauth/ -> internal/broker/ [2/8] (#648)
* refactor(broker): rename internal/oauth/ -> internal/broker/ [2/8]

Pure move. Package rename only, no behavior change.

- git mv 20 .go files and 2 embedded templates from internal/oauth/ to
  internal/broker/.
- Rewrite package declaration: oauth -> broker (20 files).
- Update first line of doc.go: "Package oauth ..." -> "Package broker ...".
- Update 3 import sites:
  * internal/aggregator/manager.go: import + 7 symbol references.
  * internal/aggregator/server.go: import + 1 symbol reference
    (musteroauth.DecodeEncryptionKey -> broker.DecodeEncryptionKey;
    external mcp-oauth aliased as `oauth` stays unaliased).
  * internal/server/oauth_http.go: import + 2 symbol references.
    (This file moves to internal/broker/http/ in PR 3.)
- Fix one live doc reference in pkg/oauth/doc.go pointing at internal/oauth.

ADRs under docs/explanation/decisions/ and the historical CHANGELOG entry
mentioning internal/oauth are intentionally left as-is (immutable history).

No back-compat shim left in internal/oauth/ -- only two aggregator files
imported it, both updated inline. This introduces a transient direct
internal/aggregator/ -> internal/broker/ import that PR 4-5 removes when
the local adapter takes over; PR 8 enforces the seam with a CI rule.

Diff: +33/-33 lines (git detected the renames). ~3050 LOC moved.
make test (-race) and golangci-lint clean across all packages.

* review: pin TLS 1.2 minimum on broker's custom-CA HTTP client

gosec G402 fires when tls.Config has no MinVersion pin. The cleanup PR
adds the same line to internal/oauth/manager.go; once that merges and
this branch rebases, the rename brings it forward and this commit
becomes redundant. Kept here so PR 648 CI passes independently of
cleanup PR merge order.

* review: refresh broker/doc.go for rename rationale + correct pkg/oauth NewClient example

internal/broker/doc.go now opens by acknowledging the package was
renamed from 'oauth'/'OAuth 2.1 proxy' to 'broker' and why (token-flow
coordination beyond pure OAuth proxy). pkg/oauth/doc.go's usage example
showed oauth.NewClient(httpClient, logger), which has never matched the
functional-option signature; fix to oauth.NewClient(oauth.WithHTTPClient(httpClient)).

* review: trim historical narrative from broker doc.go
