+++ internal/broker/api_adapter.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import (
 	"context"
