+++ internal/broker/encryption_key.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import (
 	"fmt"
