+++ internal/broker/handler.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import (
 	"bytes"
