+++ internal/broker/state_store_test.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import (
 	"encoding/base64"
