+++ internal/broker/token_store.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import (
 	"sync"
