+++ internal/broker/types.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import (
 	"time"
