+++ internal/broker/valkey_encryption.go
@@ -1,4 +1,4 @@
-package oauth
+package broker
 
 import "github.com/giantswarm/mcp-oauth/security"
