+++ pkg/oauth/doc.go
@@ -3,7 +3,7 @@
 //
 // This package contains the common OAuth functionality that is shared between
 // the agent-side OAuth implementation (internal/agent/oauth) and the server-side
-// OAuth implementation (internal/oauth).
+// OAuth implementation (internal/broker).
 //
 // # Core Components
 //
@@ -29,6 +29,6 @@
 //
 //	import "github.com/giantswarm/muster/pkg/oauth"
 //
-//	client := oauth.NewClient(httpClient, logger)
+//	client := oauth.NewClient(oauth.WithHTTPClient(httpClient))
 //	metadata, err := client.DiscoverMetadata(ctx, issuer)
 package oauth
